Gas price oracle: ignore gas price rounding for values less than 0.01
