+++ apps/explorer/lib/explorer/chain/cache/gas_price_oracle.ex
@@ -5,6 +5,18 @@ defmodule Explorer.Chain.Cache.GasPriceOracle do
 
   require Logger
 
+  import Ecto.Query,
+    only: [
+      from: 2
+    ]
+
+  alias Explorer.Chain.{
+    Block,
+    Wei
+  }
+
+  alias Explorer.Repo
+
   @default_cache_period :timer.seconds(30)
 
   @num_of_blocks (case Integer.parse(System.get_env("GAS_PRICE_ORACLE_NUM_OF_BLOCKS", "200")) do
@@ -35,7 +47,43 @@ defmodule Explorer.Chain.Cache.GasPriceOracle do
     ttl_check_interval: :timer.minutes(5),
     callback: &async_task_on_deletion(&1)
 
-  alias Explorer.Chain
+  def get_average_gas_price(num_of_blocks, safelow_percentile, average_percentile, fast_percentile) do
+    lates_gas_price_query =
+      from(
+        block in Block,
+        left_join: transaction in assoc(block, :transactions),
+        where: block.consensus == true,
+        where: transaction.status == ^1,
+        where: transaction.gas_price > ^0,
+        group_by: block.number,
+        order_by: [desc: block.number],
+        select: min(transaction.gas_price),
+        limit: ^num_of_blocks
+      )
+
+    latest_gas_prices =
+      lates_gas_price_query
+      |> Repo.all(timeout: :infinity)
+
+    latest_ordered_gas_prices =
+      latest_gas_prices
+      |> Enum.map(fn %Wei{value: gas_price} -> Decimal.to_integer(gas_price) end)
+
+    safelow_gas_price = gas_price_percentile_to_gwei(latest_ordered_gas_prices, safelow_percentile)
+    average_gas_price = gas_price_percentile_to_gwei(latest_ordered_gas_prices, average_percentile)
+    fast_gas_price = gas_price_percentile_to_gwei(latest_ordered_gas_prices, fast_percentile)
+
+    gas_prices = %{
+      "slow" => safelow_gas_price,
+      "average" => average_gas_price,
+      "fast" => fast_gas_price
+    }
+
+    {:ok, gas_prices}
+  catch
+    error ->
+      {:error, error}
+  end
 
   defp handle_fallback(:gas_prices) do
     # This will get the task PID if one exists and launch a new task if not
@@ -51,7 +99,7 @@ defmodule Explorer.Chain.Cache.GasPriceOracle do
     {:ok, task} =
       Task.start(fn ->
         try do
-          result = Chain.get_average_gas_price(@num_of_blocks, @safelow, @average, @fast)
+          result = get_average_gas_price(@num_of_blocks, @safelow, @average, @fast)
 
           set_all(result)
         rescue
@@ -65,6 +113,40 @@ defmodule Explorer.Chain.Cache.GasPriceOracle do
       end)
 
     {:update, task}
+  end
+
+  defp gas_price_percentile_to_gwei(gas_prices, percentile) do
+    gas_price_wei = percentile(gas_prices, percentile)
+
+    if gas_price_wei do
+      gas_price_gwei = Wei.to(%Wei{value: Decimal.from_float(gas_price_wei)}, :gwei)
+
+      gas_price_gwei_float = gas_price_gwei |> Decimal.to_float()
+
+      if gas_price_gwei_float > 0.01 do
+        gas_price_gwei_float
+        |> Float.ceil(2)
+      else
+        gas_price_gwei_float
+      end
+    else
+      nil
+    end
+  end
+
+  @spec percentile(list, number) :: number | nil
+  defp percentile([], _), do: nil
+  defp percentile([x], _), do: x
+  defp percentile(list, 0), do: Enum.min(list)
+  defp percentile(list, 100), do: Enum.max(list)
+
+  defp percentile(list, n) when is_list(list) and is_number(n) do
+    s = Enum.sort(list)
+    r = n / 100.0 * (length(list) - 1)
+    f = :erlang.trunc(r)
+    lower = Enum.at(s, f)
+    upper = Enum.at(s, f + 1)
+    lower + (upper - lower) * (r - f)
   end
 
   # By setting this as a `callback` an async task will be started each time the
